Collapse vault items tables into one mode-driven table
Replace the two table components and the per-column component kit with a
single VaultItemsTableComponent that renders every column inline and is
configured by a `mode` input ("individual" | "organization"):

- mode derives both the visible columns and the available bulk/per-row
  actions, so context-inappropriate actions are unrepresentable (no free
  show* flags).
- columns are inline `<bit-column>`s using the typed `dataSource().columns.x`
  directly — no `[ds]` plumbing and `let row` stays precisely typed.
- one flat VaultRow view-model (superset of the previous per-context types);
  per-row capability flags + permission text stay consumer-resolved.

Net: the feature is four files (model, table, stories, sort spec) instead of
twelve, with one cohesive component in place of two tables + a column kit.
